requirements.txt
Django==6.0.3
Pillow==10.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
